requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
backcall==0.1.0
bottle==0.12.18
bottle-cors==0.1.5
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
filelock==3.0.12
gunicorn==20.0.4
httpie==2.1.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
Pygments==2.6.1
pylint==2.5.3
pystache==0.5.4
python3-openid==3.2.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
sentry-sdk==0.15.1
six==1.15.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
toml==0.10.1
trello==0.9.4
urllib3==1.25.9
virtualenv==20.0.25
waitress==1.4.3
whitenoise==5.1.0
wrapt==1.12.1
